lark-token-generator/requirements.txt
functions-framework==3.5.0
Flask==2.3.3
requests>=2.28.0
google-cloud-secret-manager>=2.16.0
